--- a/benchmark/array_sample.yml
+++ b/benchmark/array_sample.yml
@@ -1,0 +1,5 @@
+
+prelude: ary = (1..10_000).to_a
+benchmark:
+  - ary.sample
+  - ary.sample(2)